Update README.md
Update readme to include the projects that the research team is working on during the week of 2/3/2020. The tasks include designing the lab study on mind perception, setting up a GitHub repository to hold all computational and fundamental research, progressing towards a working prototype of Knob, and progress on Tiana's ML scripts.